Struct massa_consensus_worker::controller::ConsensusControllerImpl

pub struct ConsensusControllerImpl {
    command_sender: SyncSender<ConsensusCommand>,
    channels: ConsensusChannels,
    shared_state: Arc<RwLock<ConsensusState>>,
    bootstrap_part_size: u64,
    broadcast_enabled: bool,
}

The retrieval of data is made using a shared state and modifications are asked by sending message to a channel. This is done mostly to be able to:

send commands through the channel without waiting for them to be processed from the point of view of the sending thread, and channels are very much optimal for that (much faster than locks)

still be able to read the current state of the graph as processed so far (for this we need a shared state)

Note that sending commands and reading the state is done from different, mutually-asynchronous tasks and they can have data that are not sync yet.

Fields
command_sender: SyncSender<ConsensusCommand>channels: ConsensusChannelsshared_state: Arc<RwLock<ConsensusState>>bootstrap_part_size: u64broadcast_enabled: bool
Implementations

impl ConsensusControllerImpl

pub fn new( command_sender: SyncSender<ConsensusCommand>, channels: ConsensusChannels, shared_state: Arc<RwLock<ConsensusState>>, bootstrap_part_size: u64, broadcast_enabled: bool ) -> Self